[System]: Revert my latest WebResponseStream changes.
This reverts PR #6662 (commit 18a1a25) and #6771 (commit da03780).

These two changes introduced a regression:
https://github.com/mono/mono/issues/6788

(cherry picked from commit 8bfcb2714708d7f4d5a88bcdb0e133069c998f9a)
