+++ mcs/class/System/System.Net/HttpWebResponse.cs
@@ -36,6 +36,7 @@ using System;
 using System.Collections;
 using System.Globalization;
 using System.IO;
+using System.IO.Compression;
 using System.Net.Sockets;
 using System.Runtime.Serialization;
 using System.Threading;
@@ -96,6 +97,16 @@ namespace System.Net
 			if (container != null) {
 				this.cookie_container = container;	
 				FillCookies ();
+			}
+
+			string content_encoding = webHeaders ["Content-Encoding"];
+			if (content_encoding == "gzip" && (stream.Request.AutomaticDecompression & DecompressionMethods.GZip) != 0) {
+				this.stream = new GZipStream (stream, CompressionMode.Decompress);
+				webHeaders.Remove (HttpRequestHeader.ContentEncoding);
+			}
+			else if (content_encoding == "deflate" && (stream.Request.AutomaticDecompression & DecompressionMethods.Deflate) != 0) {
+				this.stream = new DeflateStream (stream, CompressionMode.Decompress);
+				webHeaders.Remove (HttpRequestHeader.ContentEncoding);
 			}
 		}
