+++ mcs/class/System/System.Net/WebResponseStream.cs
@@ -24,7 +24,6 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 using System.IO;
-using System.IO.Compression;
 using System.Text;
 using System.Collections;
 using System.Collections.Generic;
@@ -37,7 +36,7 @@ namespace System.Net
 {
 	class WebResponseStream : WebConnectionStream
 	{
-		Stream innerStreamWrapper;
+		BufferOffsetSize readBuffer;
 		long contentLength;
 		long totalRead;
 		bool nextReadCalled;
@@ -99,8 +98,14 @@ namespace System.Net
 
 		public override bool CanWrite => false;
 
-		bool ChunkedRead {
-			get; set;
+		protected bool ChunkedRead {
+			get;
+			private set;
+		}
+
+		protected MonoChunkStream ChunkStream {
+			get;
+			private set;
 		}
 
 		public override async Task<int> ReadAsync (byte[] buffer, int offset, int size, CancellationToken cancellationToken)
@@ -135,12 +140,12 @@ namespace System.Net
 
 			WebConnection.Debug ($"{ME} READ ASYNC #2: {totalRead} {contentLength}");
 
-			int nbytes = 0;
+			int oldBytes = 0, nbytes = 0;
 			Exception throwMe = null;
 
 			try {
 				// FIXME: NetworkStream.ReadAsync() does not support cancellation.
-				nbytes = await HttpWebRequest.RunWithTimeout (
+				(oldBytes, nbytes) = await HttpWebRequest.RunWithTimeout (
 					ct => ProcessRead (buffer, offset, size, ct),
 					ReadTimeout, () => {
 						Operation.Abort ();
@@ -150,7 +155,7 @@ namespace System.Net
 				throwMe = GetReadException (WebExceptionStatus.ReceiveFailure, e, "ReadAsync");
 			}
 
-			WebConnection.Debug ($"{ME} READ ASYNC #3: {totalRead} {contentLength} - {nbytes} {throwMe?.Message}");
+			WebConnection.Debug ($"{ME} READ ASYNC #3: {totalRead} {contentLength} - {oldBytes} {nbytes} {throwMe?.Message}");
 
 			if (throwMe != null) {
 				lock (locker) {
@@ -165,65 +170,133 @@ namespace System.Net
 			}
 
 			lock (locker) {
-				readTcs.TrySetResult (nbytes);
+				readTcs.TrySetResult (oldBytes + nbytes);
 				readTcs = null;
 				nestedRead = 0;
 			}
 
 			if (totalRead >= contentLength && !nextReadCalled) {
-				WebConnection.Debug ($"{ME} READ ASYNC - READ COMPLETE: {nbytes} - {totalRead} {contentLength} {nextReadCalled}");
+				WebConnection.Debug ($"{ME} READ ASYNC - READ COMPLETE: {oldBytes} {nbytes} - {totalRead} {contentLength} {nextReadCalled}");
 				if (!nextReadCalled) {
 					nextReadCalled = true;
 					Operation.CompleteResponseRead (true);
 				}
 			}
 
-			return nbytes;
+			return oldBytes + nbytes;
 		}
 
-		async Task<int> ProcessRead (byte[] buffer, int offset, int size, CancellationToken cancellationToken)
+		async Task<(int, int)> ProcessRead (byte[] buffer, int offset, int size, CancellationToken cancellationToken)
 		{
 			WebConnection.Debug ($"{ME} PROCESS READ: {totalRead} {contentLength}");
 
 			cancellationToken.ThrowIfCancellationRequested ();
-			if (read_eof || totalRead >= contentLength) {
+			if (totalRead >= contentLength) {
 				read_eof = true;
 				contentLength = totalRead;
-				return 0;
+				return (0, 0);
+			}
+
+			int oldBytes = 0;
+			int remaining = readBuffer?.Size ?? 0;
+			if (remaining > 0) {
+				int copy = (remaining > size) ? size : remaining;
+				Buffer.BlockCopy (readBuffer.Buffer, readBuffer.Offset, buffer, offset, copy);
+				readBuffer.Offset += copy;
+				readBuffer.Size -= copy;
+				offset += copy;
+				size -= copy;
+				totalRead += copy;
+				if (totalRead >= contentLength) {
+					contentLength = totalRead;
+					read_eof = true;
+				}
+				if (size == 0 || totalRead >= contentLength)
+					return (0, copy);
+				oldBytes = copy;
 			}
 
 			if (contentLength != Int64.MaxValue && contentLength - totalRead < size)
 				size = (int)(contentLength - totalRead);
 
-			WebConnection.Debug ($"{ME} PROCESS READ #1: {size} {read_eof}");
+			WebConnection.Debug ($"{ME} PROCESS READ #1: {oldBytes} {size} {read_eof}");
+
+			if (read_eof) {
+				contentLength = totalRead;
+				return (oldBytes, 0);
+			}
 
 			var ret = await InnerReadAsync (buffer, offset, size, cancellationToken).ConfigureAwait (false);
 
 			if (ret <= 0) {
 				read_eof = true;
 				contentLength = totalRead;
-				return ret;
+				return (oldBytes, 0);
 			}
 
 			totalRead += ret;
-			return ret;
+			return (oldBytes, ret);
 		}
 
-		async Task<int> InnerReadAsync (byte[] buffer, int offset, int size, CancellationToken cancellationToken)
+		internal async Task<int> InnerReadAsync (byte[] buffer, int offset, int size, CancellationToken cancellationToken)
 		{
+			WebConnection.Debug ($"{ME} INNER READ ASYNC");
+
 			Operation.ThrowIfDisposed (cancellationToken);
 
-			WebConnection.Debug ($"{ME} INNER READ ASYNC: stream={innerStreamWrapper}");
+			int nbytes = 0;
+			bool done = false;
+
+			if (!ChunkedRead || (!ChunkStream.DataAvailable && ChunkStream.WantMore)) {
+				nbytes = await InnerStream.ReadAsync (buffer, offset, size, cancellationToken).ConfigureAwait (false);
+				WebConnection.Debug ($"{ME} INNER READ ASYNC #1: {nbytes} {ChunkedRead}");
+				if (!ChunkedRead)
+					return nbytes;
+				done = nbytes == 0;
+			}
 
 			try {
-				var ret = await innerStreamWrapper.ReadAsync (
-					buffer, offset, size, cancellationToken).ConfigureAwait (false);
-				WebConnection.Debug ($"{ME} INNER READ ASYNC DONE: {ret}");
-				return ret;
-			} catch (Exception ex) {
-				WebConnection.Debug ($"{ME} INNER READ ASYNC EX: {ex.Message}");
-				throw;
+				ChunkStream.WriteAndReadBack (buffer, offset, size, ref nbytes);
+				WebConnection.Debug ($"{ME} INNER READ ASYNC #1: {done} {nbytes} {ChunkStream.WantMore}");
+				if (!done && nbytes == 0 && ChunkStream.WantMore)
+					nbytes = await EnsureReadAsync (buffer, offset, size, cancellationToken).ConfigureAwait (false);
+			} catch (Exception e) {
+				if (e is WebException || e is OperationCanceledException)
+					throw;
+				throw new WebException ("Invalid chunked data.", e, WebExceptionStatus.ServerProtocolViolation, null);
 			}
+
+			if ((done || nbytes == 0) && ChunkStream.ChunkLeft != 0) {
+				// HandleError (WebExceptionStatus.ReceiveFailure, null, "chunked EndRead");
+				throw new WebException ("Read error", null, WebExceptionStatus.ReceiveFailure, null);
+			}
+
+			return nbytes;
+		}
+
+		async Task<int> EnsureReadAsync (byte[] buffer, int offset, int size, CancellationToken cancellationToken)
+		{
+			byte[] morebytes = null;
+			int nbytes = 0;
+			while (nbytes == 0 && ChunkStream.WantMore && !cancellationToken.IsCancellationRequested) {
+				int localsize = ChunkStream.ChunkLeft;
+				if (localsize <= 0) // not read chunk size yet
+					localsize = 1024;
+				else if (localsize > 16384)
+					localsize = 16384;
+
+				if (morebytes == null || morebytes.Length < localsize)
+					morebytes = new byte[localsize];
+
+				int nread = await InnerStream.ReadAsync (morebytes, 0, localsize, cancellationToken).ConfigureAwait (false);
+				if (nread <= 0)
+					return 0; // Error
+
+				ChunkStream.Write (morebytes, 0, nread);
+				nbytes += ChunkStream.Read (buffer, offset + nbytes, size - nbytes);
+			}
+
+			return nbytes;
 		}
 
 		bool CheckAuthHeader (string headerName)
@@ -248,7 +321,7 @@ namespace System.Net
 			}
 		}
 
-		void Initialize (BufferOffsetSize buffer)
+		async Task Initialize (BufferOffsetSize buffer, CancellationToken cancellationToken)
 		{
 			WebConnection.Debug ($"{ME} INIT: status={(int)StatusCode} bos={buffer.Offset}/{buffer.Size}");
 
@@ -277,28 +350,35 @@ namespace System.Net
 			if (!Int32.TryParse (clength, out stream_length))
 				stream_length = -1;
 
+			string me = "WebResponseStream.Initialize()";
 			string tencoding = null;
 			if (ExpectContent)
 				tencoding = Headers["Transfer-Encoding"];
 
 			ChunkedRead = (tencoding != null && tencoding.IndexOf ("chunked", StringComparison.OrdinalIgnoreCase) != -1);
-
-			if (ChunkedRead) {
-				innerStreamWrapper = new MonoChunkStream (
-					Operation, CreateStreamWrapper (buffer), Headers);
-			} else if (!IsNtlmAuth () && contentLength > 0 && buffer.Size >= contentLength) {
-				innerStreamWrapper = new BufferedReadStream (Operation, null, buffer);
+			if (!ChunkedRead) {
+				readBuffer = buffer;
+				try {
+					if (contentLength > 0 && readBuffer.Size >= contentLength) {
+						if (!IsNtlmAuth ())
+							await ReadAllAsync (false, cancellationToken).ConfigureAwait (false);
+					}
+				} catch (Exception e) {
+					throw GetReadException (WebExceptionStatus.ReceiveFailure, e, me);
+				}
+			} else if (ChunkStream == null) {
+				try {
+					ChunkStream = new MonoChunkStream (buffer.Buffer, buffer.Offset, buffer.Offset + buffer.Size, Headers);
+				} catch (Exception e) {
+					throw GetReadException (WebExceptionStatus.ServerProtocolViolation, e, me);
+				}
 			} else {
-				innerStreamWrapper = CreateStreamWrapper (buffer);
-			}
-
-			string content_encoding = Headers["Content-Encoding"];
-			if (content_encoding == "gzip" && (Request.AutomaticDecompression & DecompressionMethods.GZip) != 0) {
-				innerStreamWrapper = new GZipStream (innerStreamWrapper, CompressionMode.Decompress);
-				Headers.Remove (HttpRequestHeader.ContentEncoding);
-			} else if (content_encoding == "deflate" && (Request.AutomaticDecompression & DecompressionMethods.Deflate) != 0) {
-				innerStreamWrapper = new DeflateStream (innerStreamWrapper, CompressionMode.Decompress);
-				Headers.Remove (HttpRequestHeader.ContentEncoding);
+				ChunkStream.ResetBuffer ();
+				try {
+					ChunkStream.Write (buffer.Buffer, buffer.Offset, buffer.Size);
+				} catch (Exception e) {
+					throw GetReadException (WebExceptionStatus.ServerProtocolViolation, e, me);
+				}
 			}
 
 			WebConnection.Debug ($"{ME} INIT #1: - {ExpectContent} {closed} {nextReadCalled}");
@@ -311,13 +391,6 @@ namespace System.Net
 				}
 				Operation.CompleteResponseRead (true);
 			}
-		}
-
-		Stream CreateStreamWrapper (BufferOffsetSize buffer)
-		{
-			if (buffer == null || buffer.Size == 0)
-				return InnerStream;
-			return new BufferedReadStream (Operation, InnerStream, buffer);
 		}
 
 		internal async Task ReadAllAsync (bool resending, CancellationToken cancellationToken)
@@ -357,6 +430,7 @@ namespace System.Net
 				if (totalRead >= contentLength)
 					return;
 
+				byte[] b = null;
 				int new_size;
 
 				if (contentLength == Int64.MaxValue && !ChunkedRead) {
@@ -377,10 +451,19 @@ namespace System.Net
 					KeepAlive = false;
 				}
 
-				BufferOffsetSize bos;
 				if (contentLength == Int64.MaxValue) {
-					var ms = new MemoryStream ();
-					var buffer = new BufferOffsetSize (new byte[8192], false);
+					MemoryStream ms = new MemoryStream ();
+					BufferOffsetSize buffer = null;
+					if (readBuffer != null && readBuffer.Size > 0) {
+						ms.Write (readBuffer.Buffer, readBuffer.Offset, readBuffer.Size);
+						readBuffer.Offset = 0;
+						readBuffer.Size = readBuffer.Buffer.Length;
+						if (readBuffer.Buffer.Length >= 8192)
+							buffer = readBuffer;
+					}
+
+					if (buffer == null)
+						buffer = new BufferOffsetSize (new byte[8192], false);
 
 					int read;
 					while ((read = await InnerReadAsync (buffer.Buffer, buffer.Offset, buffer.Size, cancellationToken)) != 0)
@@ -388,23 +471,29 @@ namespace System.Net
 
 					new_size = (int)ms.Length;
 					contentLength = new_size;
-					bos = new BufferOffsetSize (ms.GetBuffer (), 0, new_size, false);
+					readBuffer = new BufferOffsetSize (ms.GetBuffer (), 0, new_size, false);
 				} else {
 					new_size = (int)(contentLength - totalRead);
-					var b = new byte[new_size];
+					b = new byte[new_size];
 					int readSize = 0;
-					int remaining = new_size;
+					if (readBuffer != null && readBuffer.Size > 0) {
+						readSize = readBuffer.Size;
+						if (readSize > new_size)
+							readSize = new_size;
+
+						Buffer.BlockCopy (readBuffer.Buffer, readBuffer.Offset, b, 0, readSize);
+					}
+
+					int remaining = new_size - readSize;
 					int r = -1;
 					while (remaining > 0 && r != 0) {
-						r = await InnerReadAsync (b, readSize, remaining, cancellationToken).ConfigureAwait (false);
+						r = await InnerReadAsync (b, readSize, remaining, cancellationToken);
 						remaining -= r;
 						readSize += r;
 					}
-					bos = new BufferOffsetSize (b, 0, new_size, false);
 				}
 
-				innerStreamWrapper = new BufferedReadStream (Operation, null, bos);
-
+				readBuffer = new BufferOffsetSize (b, 0, new_size, false);
 				totalRead = 0;
 				nextReadCalled = true;
 				myReadTcs.TrySetResult (new_size);
@@ -516,7 +605,8 @@ namespace System.Net
 			WebConnection.Debug ($"{ME} INIT READ ASYNC LOOP DONE: {buffer.Offset} {buffer.Size}");
 
 			try {
-				Initialize (buffer);
+				Operation.ThrowIfDisposed (cancellationToken);
+				await Initialize (buffer, cancellationToken).ConfigureAwait (false);
 			} catch (Exception e) {
 				throw GetReadException (WebExceptionStatus.ReceiveFailure, e, "ReadDoneAsync6");
 			}
@@ -637,5 +727,7 @@ namespace System.Net
 
 			throw GetReadException (WebExceptionStatus.ServerProtocolViolation, null, "GetResponse");
 		}
+
+
 	}
 }
