+++ mcs/class/System/common_networking.sources
@@ -49,9 +49,6 @@ System.Net/WebConnectionTunnel.cs
 System.Net/WebOperation.cs
 System.Net/WebRequestStream.cs
 System.Net/WebResponseStream.cs
-System.Net/MonoChunkStream.cs
-System.Net/WebReadStream.cs
-System.Net/BufferedReadStream.cs
 
 ../referencesource/System/net/System/Net/Sockets/TCPClient.cs
 ../referencesource/System/net/System/Net/Sockets/TCPListener.cs
